+++ src/prime_rl/orchestrator/orchestrator.py
@@ -39,8 +39,9 @@ from prime_rl.utils.client import (
     init_nccl_broadcast,
     reload_weights,
     setup_admin_clients,
-    setup_clients,
     setup_evals_client,
+    setup_threaded_clients,
+    teardown_clients,
     update_weights,
 )
 from prime_rl.utils.heartbeat import Heartbeat
@@ -93,9 +94,9 @@ async def orchestrate(config: OrchestratorConfig):
 
     # Setup client
     logger.info(
-        f"Initializing OpenAI client (base_url={', '.join(config.client.base_url)}, api_key_var={config.client.api_key_var}, headers={config.client.headers})"
+        f"Initializing OpenAI client (base_url={', '.join(config.client.base_url)}, api_key_var={config.client.api_key_var}, headers={config.client.headers}, max_workers_per_client={config.client.max_workers_per_client})"
     )
-    clients = setup_clients(config.client)
+    clients = setup_threaded_clients(config.client)
     admin_clients = setup_admin_clients(config.client)
     evals_client = setup_evals_client()
 
@@ -525,6 +526,9 @@ async def orchestrate(config: OrchestratorConfig):
 
     # Cancel event loop lag monitor task
     event_loop_lag_monitor_task.cancel()
+
+    # Teardown clients
+    teardown_clients(clients)
 
     logger.success("Orchestrator finished.")